Refactor: NoteCombine관련 로직 분리
